# lib/api-zod/src/generated/types/oppFeed.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OppOpportunity } from './oppOpportunity';

export interface OppFeed {
  generatedAt: string;
  userState: string;
  userRegion: string;
  summary: string;
  opportunities: OppOpportunity[];
  linkedinTips: string[];
  cached: boolean;
  cacheAge?: number;
}