# lib/api-zod/src/generated/types/oppLoginInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OppLoginInput {
  /** @minLength 2 */
  name: string;
  /** @minLength 5 */
  email: string;
  /** @minLength 2 */
  state: string;
  /** @minLength 2 */
  region: string;
}