# lib/api-zod/src/generated/types/oppOpportunity.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OppOpportunity {
  id: string;
  title: string;
  description: string;
  category: string;
  platform: string;
  link: string;
  deadline: string;
  location: string;
  isRemote: boolean;
  tags: string[];
}