# lib/api-zod/src/generated/types/oppUser.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OppUser {
  id: string;
  name: string;
  email: string;
  state: string;
  region: string;
}